web/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-db-logger==0.1.12
pyyaml==6.0
uritemplate==4.1.1
Markdown==3.3.7